+++ packages/executor/src/utils/index.ts
@@ -1,4 +1,3 @@
-
-export * from './bytes32toRuneId';
-
-export * from './runeIdToBytes32';
+export * from "./bytes32toRuneId";
+export * from "./getEVMAddress";
+export * from "./runeIdToBytes32";