+++ apps/docs/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @midl-xyz/midl-js-docs
 
+## 0.0.24
+
+### Patch Changes
+
+- Updated dependencies
+  - @midl-xyz/midl-js-core@0.0.21
+  - @midl-xyz/midl-js-react@0.0.25
+
 ## 0.0.23
 
 ### Patch Changes