+++ packages/executor/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @midl-xyz/midl-js-executor
 
+## 0.0.28
+
+### Patch Changes
+
+- fix: evm address extraction
+- Updated dependencies
+  - @midl-xyz/midl-js-core@0.0.21
+  - @midl-xyz/midl-js-react@0.0.25
+
 ## 0.0.27
 
 ### Patch Changes